Remove top-level require when library is loaded
I originally thought it would be desirable to have the lib automatically
load the RSpec integration whenever it's required. However in practice
it seems that under certain circumstances this library might be loaded
before RSpec "boots". Hmm not sure what that means, but the symptom has
been that RSpec::Matchers isn't defined at the time of load. So the
solution here is to have the user explicitly require `act_like/rspec` to
define the customer matcher. It's assumed that this would be done in a
test helper file.